<REQ><PDFCCE-6953><T610R_DEV_V2.0:SPCSS00870978共性问题review评审，两个稳定性问题请帮忙提供sample code>
(cherry picked from commit 00540cb23534cf2c0a18697160a8597ea9e7ccaf)

+++ kernel/exit.c
@@ -784,7 +784,7 @@ void __noreturn do_exit(long code)
 		panic("Aiee, killing interrupt handler!");
 	if (unlikely(!tsk->pid))
 		panic("Attempted to kill the idle task!");
-	if (unlikely(tsk->pid == 1))
+	if (unlikely(tsk->pid == 1) && (code != 0))
 		panic("Attempted to kill the init task!");
 
 	/*
